+++ esp_h264/test_apps/main/test_case.c
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2024 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2024-2025 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -65,6 +65,32 @@ TEST_CASE("hw_enc_single_hw_enc_qp_test", "[esp_h264]")
         TEST_ASSERT_EQUAL(ESP_H264_ERR_OK, single_hw_enc_process(cfg));
     }
 }
+
+#if CONFIG_ESP_REV_MIN_FULL >= 300
+TEST_CASE("hw_enc_single_hw_enc_pic_type_test", "[esp_h264]")
+{
+    esp_h264_raw_format_t pic_type_list[] = {
+        ESP_H264_RAW_FMT_BGR888,
+        ESP_H264_RAW_FMT_BGR565_BE,
+        ESP_H264_RAW_FMT_VUY,
+        ESP_H264_RAW_FMT_UYVY,
+        ESP_H264_RAW_FMT_O_UYY_E_VYY,
+    };
+
+    for (int16_t pic_type = 0; pic_type < sizeof(pic_type_list) / sizeof(pic_type_list[0]); pic_type++) {
+        esp_h264_enc_cfg_hw_t cfg = { 0 };
+        cfg.gop = 5;
+        cfg.fps = 30;
+        cfg.res.width = res_width;
+        cfg.res.height = res_height;
+        cfg.rc.bitrate = cfg.res.width * cfg.res.height * cfg.fps / 20;
+        cfg.rc.qp_min = 26;
+        cfg.rc.qp_max = 26;
+        cfg.pic_type = pic_type_list[pic_type];
+        TEST_ASSERT_EQUAL(ESP_H264_ERR_OK, single_hw_enc_process(cfg));
+    }
+}
+#endif
 
 TEST_CASE("hw_enc_dual_hw_enc_gop_test", "[esp_h264]")
 {
@@ -140,6 +166,43 @@ TEST_CASE("hw_enc_dual_hw_enc_qp_test", "[esp_h264]")
         TEST_ASSERT_EQUAL(ESP_H264_ERR_OK, dual_hw_enc_process(cfg));
     }
 }
+
+#if CONFIG_ESP_REV_MIN_FULL >= 300
+TEST_CASE("hw_enc_dual_hw_enc_pic_type_test", "[esp_h264]")
+{
+    esp_h264_raw_format_t pic_type_list[] = {
+        ESP_H264_RAW_FMT_BGR888,
+        ESP_H264_RAW_FMT_BGR565_BE,
+        ESP_H264_RAW_FMT_VUY,
+        ESP_H264_RAW_FMT_UYVY,
+        ESP_H264_RAW_FMT_O_UYY_E_VYY,
+    };
+    esp_h264_enc_cfg_dual_hw_t cfg;
+    for (int16_t pic_type = 0; pic_type < sizeof(pic_type_list) / sizeof(pic_type_list[0]); pic_type++) {
+        cfg.cfg0.gop = 5;
+        cfg.cfg0.fps = 30;
+        cfg.cfg0.res.width = res_width;
+        cfg.cfg0.res.height = res_height;
+        cfg.cfg0.rc.bitrate = cfg.cfg0.res.width * cfg.cfg0.res.height * cfg.cfg0.fps / 20;
+        cfg.cfg0.rc.qp_min = 26;
+        cfg.cfg0.rc.qp_max = (cfg.cfg0.rc.qp_min + 1) > 51 ? 51 : (cfg.cfg0.rc.qp_min + 1);
+        cfg.cfg0.pic_type = pic_type_list[pic_type];
+
+        cfg.cfg1.gop = 5;
+        cfg.cfg1.fps = 30;
+        cfg.cfg1.res.width = res_width1;
+        cfg.cfg1.res.height = res_height1;
+        cfg.cfg1.rc.bitrate = cfg.cfg1.res.width * cfg.cfg1.res.height * cfg.cfg1.fps / 20;
+        cfg.cfg1.rc.qp_min = 26;
+        cfg.cfg1.rc.qp_max = (cfg.cfg1.rc.qp_min + 1) > 51 ? 51 : (cfg.cfg1.rc.qp_min + 1);
+        cfg.cfg1.pic_type = pic_type_list[pic_type];
+        TEST_ASSERT_EQUAL(ESP_H264_ERR_OK, dual_hw_enc_process(cfg));
+    }
+    cfg.cfg0.pic_type = pic_type_list[0];
+    cfg.cfg1.pic_type = pic_type_list[1];
+    TEST_ASSERT_EQUAL(ESP_H264_ERR_OK, dual_hw_enc_process(cfg));
+}
+#endif
 
 TEST_CASE("hw_enc_set_get_param_single_hw_enc_thread_test", "[esp_h264]")
 {